refactor: :recycle: require login to accept invite
